Rearrange methods for filter search so that searches are partially initialized in SQL mode (no search on column values)
