+++ backend/src/main/java/com/bakdata/conquery/models/worker/LocalNamespace.java
@@ -1,11 +1,14 @@
 package com.bakdata.conquery.models.worker;
 
 import java.util.List;
+import java.util.Set;
 
 import com.bakdata.conquery.io.jackson.Injectable;
 import com.bakdata.conquery.io.storage.NamespaceStorage;
+import com.bakdata.conquery.models.datasets.Column;
 import com.bakdata.conquery.models.index.IndexService;
 import com.bakdata.conquery.models.jobs.JobManager;
+import com.bakdata.conquery.models.jobs.UpdateFilterSearchJob;
 import com.bakdata.conquery.models.query.ExecutionManager;
 import com.bakdata.conquery.models.query.FilterSearch;
 import com.bakdata.conquery.sql.execution.SqlExecutionService;
@@ -33,7 +36,22 @@ public class LocalNamespace extends Namespace {
 	}
 
 	@Override
+	void updateFilterSearch() {
+		getJobManager().addSlowJob(new UpdateFilterSearchJob(this, getFilterSearch().getIndexConfig(), this::buildSearchForColumnValuesAsync));
+	}
+
+	@Override
+	void updateMatchingStats() {
+		// TODO Build basic statistic on data
+	}
+
+	@Override
+	void buildSearchForColumnValuesAsync(Set<Column> columns) {
+		// TODO Gather values for these columns and register them
+	}
+
+	@Override
 	public void postprocessData() {
-		// Not implemented yet
+		// TODO Not implemented yet
 	}
 }
