+++ backend/src/main/java/com/bakdata/conquery/models/worker/Namespace.java
@@ -1,13 +1,17 @@
 package com.bakdata.conquery.models.worker;
 
 import java.io.IOException;
+import java.util.Collection;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Set;
 
 import com.bakdata.conquery.io.jackson.Injectable;
 import com.bakdata.conquery.io.storage.NamespaceStorage;
+import com.bakdata.conquery.models.datasets.Column;
 import com.bakdata.conquery.models.datasets.Dataset;
 import com.bakdata.conquery.models.datasets.PreviewConfig;
+import com.bakdata.conquery.models.datasets.concepts.Searchable;
 import com.bakdata.conquery.models.datasets.concepts.select.connector.specific.MappableSingleColumnSelect;
 import com.bakdata.conquery.models.identifiable.CentralRegistry;
 import com.bakdata.conquery.models.identifiable.ids.specific.DatasetId;
@@ -122,9 +126,42 @@ public abstract class Namespace extends IdResolveContext {
 		throw new UnsupportedOperationException();
 	}
 
+
+	/**
+	 * Issues a job that initializes the search that is used by the frontend for recommendations in the filter interface of a concept.
+	 */
+	abstract void updateFilterSearch();
+
+	/**
+	 * Issues a job that collects basic metrics for every concept and its nodes. This information is displayed in the frontend.
+	 */
+	abstract void updateMatchingStats();
+
+	/**
+	 * This collects the string values of the given {@link Column}s (each is a {@link com.bakdata.conquery.models.datasets.concepts.Searchable})
+	 * and registers them in the namespace's {@link FilterSearch#registerValues(Searchable, Collection)}.
+	 * After value registration for a column is complete, {@link FilterSearch#shrinkSearch(Searchable)} should be called.
+	 *
+	 * @param columns
+	 */
+	abstract void buildSearchForColumnValuesAsync(Set<Column> columns);
+
 	/**
 	 * Hook for actions that are best done after all data has been imported and is in a consistent state.
 	 * Such actions are for example search initialization and collection of matching statistics.
+	 *
+	 * @implNote This intentionally submits a SlowJob so that it will be queued after all jobs that are already in the queue (usually import jobs).
 	 */
-	public abstract void postprocessData();
+	public void postprocessData() {
+
+		getJobManager().addSlowJob(new SimpleJob(
+				"Initiate Update Matching Stats and FilterSearch",
+				() -> {
+					updateMatchingStats();
+					updateFilterSearch();
+					updateInternToExternMappings();
+				}
+		));
+
+	}
 }
